requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.9
jinja2==3.1.4
sentence-transformers==3.1.1
faiss-cpu==1.8.0.post1
rank-bm25==0.2.2
numpy==1.26.4
scikit-learn==1.5.2
pytest==8.3.3
httpx==0.27.2
pyyaml==6.0.2
rapidfuzz==3.9.6
tf-keras
